fix(build): close the five failures the pipeline has been red on since TX-002
None of these came from the text-source work; all five predate it and none of
them is reachable from a Linux gcc build, which is why they survived. Written
down here so the next one is found the same way.

**macOS.** `ConsoleMode` remembers `c_iflag` as an `unsigned int` so its header
need not include `<termios.h>`, and asserts the two agree. `tcflag_t` is
`unsigned int` on Linux and `unsigned long` on macOS, so the assertion was
guaranteed to fail on exactly one platform in the matrix. It remembers an
`unsigned long` now, which is wide enough for both.

**Windows, both compilers.** `std::freopen` is a deprecation error under
`/W4 -Werror` on MSVC and clang-cl alike. `freopen_s` there, the plain one
everywhere else: the same in-place reopen of `stdin`, which is the whole point —
FTXUI holds that handle and anything opening a second stream would leave it
reading the drained pipe.

**The version guard.** A test fixture wrote `2.0.0-alpha.6` out twice. It reads
`kVersionString` now, which is what the guard exists to insist on: a literal in
a test is a second place the version lives, and the second place is the one that
goes stale.

**typos.** "ANDed" reads as a misspelling of "dead" to the spell checker, which
is fair enough. Both occurrences say what they meant in words instead.

**clang-tidy.** Nine `bugprone-unchecked-optional-access` findings, every one of
them the same shape: `(*result)->member` on a `Result<std::optional<T>>`. The
check cannot follow the `has_value` test through the `Result`, so it calls every
use unchecked. Naming the optional in a local is the fix this codebase already
found once — there is a comment in `store` saying so — and it is applied at the
other nine sites now. The registry's `add` takes its `shared_ptr` by const
reference for the same job's other finding.

This one is the least verifiable locally: CI runs Ubuntu's clang-tidy 18 against
libc++, and the check models libc++'s `optional` internals, so a clang 22 and
libstdc++ machine reports nothing either before or after. It is the established
pattern rather than a confirmed fix, and the pipeline is the only place that can
say so.

gcc, clang and ASan/UBSan green at 1811 tests; clang-format 18.1.8 clean; the
literal guard passes.

+++ libs/tui/src/ConsoleMode.h
@@ -71,9 +71,12 @@ namespace typeit::tui {
         /// before FTXUI ran would undo FTXUI's raw mode as a side effect, and
         /// the two objects have no idea about each other's lifetime.
         ///
-        /// `unsigned int` rather than `tcflag_t`, so the header needs no
+        /// `unsigned long` rather than `tcflag_t`, so the header needs no
         /// `<termios.h>`; the one place that uses it checks the two agree.
-        unsigned int previous_input_flags_ = 0;
+        /// Not `unsigned int`: `tcflag_t` is that on Linux and `unsigned long`
+        /// on macOS, and the narrower of the two is a static assertion failure
+        /// on exactly one of the platforms in the matrix.
+        unsigned long previous_input_flags_ = 0;
         bool captured_ = false;
 #endif
     };
